Add files via upload
Agri Sales Management System is to help farmers by providing all agriculture related information in the website. It is a farmer management website application which helps farmer to give best practice farming processes. It helps farmer to improve their productivity and profitability. It can be defined as Intelligent Agriculture. The purpose is to automate the existing manual system by the help of computerized equipment's and full-fledged computer software, fulfilling their requirements, so that their valuable data can be stored for a longer period with easy accessing and manipulation of the same. Basically, the project describes how to manage for good performance and better services for clients.